Auth: expose testing API
